+++ ChangeLog.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Partial support for transmission materials in built-in render pipeline (extension [KHR_materials_transmission](https://github.com/KhronosGroup/glTF/tree/master/extensions/2.0/Khronos/KHR_materials_transmission); see #111 for details)
 ### Changed
 - Performance improvement: Avoid redundant Shader.Find invocations by making cached shader references static
+- Built-In shaders can customized now by overriding `BuiltInMaterialGenerator.FindShader*` methods
 ### Fixed
 - Unlit double sided shader graph materials